♻️ refactor(webLotto): 코드 내용과 어울리지 않는 파일에 있던 코드 분리
